Just put all jeckle errors on the same file, no need to separate them, nor have another namespace wrapping them
